chore: 彻底解决高分屏,不同缩放比可能导致的图片不清晰问题
彻底解决高分屏,不同缩放比可能导致的图片不清晰问题

Log: 彻底解决高分屏,不同缩放比可能导致的图片不清晰问题

+++ src/document/DjVuModel.cpp
@@ -445,7 +445,7 @@ QSizeF DjVuPage::sizeF() const
     return m_size;
 }
 
-QImage DjVuPage::render(qreal width, qreal height, const QRect &slice, Qt::AspectRatioMode mode)const
+QImage DjVuPage::render(int width, int height, const QRect &slice, Qt::AspectRatioMode mode)const
 {
     LOCK_PAGE
