+++ src/document/DjVuModel.h
@@ -45,7 +45,7 @@ public:
 
     QSizeF sizeF() const override;
 
-    QImage render(qreal width, qreal height, const QRect &slice = QRect(), Qt::AspectRatioMode mode = Qt::IgnoreAspectRatio)const override;     //按大小获取
+    QImage render(int width, int height, const QRect &slice = QRect(), Qt::AspectRatioMode mode = Qt::IgnoreAspectRatio)const override;     //按大小获取
 
     QImage render(const double scaleFactor = 1.0, const QRect &slice = QRect()) const override;
